Extract common parser
