+++ src/clients/ethers/ether-deployer-sdk-write.ts
@@ -17,12 +17,7 @@ import {
   TransactionOptions,
 } from "../../types";
 import { extractEventArgument, generateSalt } from "../../utils/helper";
-import {
-  ensureProtocolFee,
-  validateFeeSplitArray,
-  validateLaunchTokenBondingCurveParams,
-  validateSalt,
-} from "../../utils/validators";
+import { processLaunchTokenParams } from "../../utils/validators";
 import type { EthersSDKConfig } from "./types";
 
 const ERC20_ABI = [
@@ -266,36 +261,7 @@ export class DeployerWriter {
     if (!this.signer)
       throw new Error("Signer is required for launching tokens");
 
-    // Adjust fee splits to ensure protocol fee is included
-    params.bondingCurveFeeSplits = ensureProtocolFee(
-      "bondingCurveFeeSplits",
-      BigInt(params.bondingCurveBuyFee) === 0n &&
-        BigInt(params.bondingCurveSellFee) === 0n,
-      params.bondingCurveFeeSplits
-    );
-    params.poolFeeSplits = ensureProtocolFee(
-      "poolFeeSplits",
-      params.poolFees === 0,
-      params.poolFeeSplits
-    );
-    params.graduationFeeSplits = ensureProtocolFee(
-      "graduationFeeSplits",
-      BigInt(params.graduationFeeBps) === 0n,
-      params.graduationFeeSplits
-    );
-    // Validate after injecting protocol fee
-    validateLaunchTokenBondingCurveParams(params);
-    validateFeeSplitArray(
-      params.bondingCurveFeeSplits,
-      "bondingCurveFeeSplits"
-    );
-    validateFeeSplitArray(params.poolFeeSplits, "poolFeeSplits");
-    validateFeeSplitArray(params.graduationFeeSplits, "graduationFeeSplits");
-
-    const config = this.buildTokenDeploymentConfig(params);
-    const finalSalt = salt ?? generateSalt();
-    validateSalt(finalSalt);
-
+    const { config, salt: finalSalt } = processLaunchTokenParams(params, salt);
     const txOverrides: any = {};
     if (options?.gasPrice) txOverrides.gasPrice = options.gasPrice;
 
@@ -450,63 +416,5 @@ export class DeployerWriter {
       console.error("Error waiting for transaction receipt:", error);
       throw error;
     }
-  }
-
-  private buildTokenDeploymentConfig(
-    params: LaunchTokenParams
-  ): TokenDeploymentConfig {
-    const totalSupply = (
-      BigInt(params.teamSupply) +
-      BigInt(params.bondingCurveSupply) +
-      BigInt(params.liquidityPoolSupply)
-    ).toString();
-
-    return {
-      creator: params.creator,
-      baseToken: params.baseToken,
-      name: params.name,
-      symbol: params.symbol,
-      image: params.image,
-      appIdentifier: "",
-      teamSupply: BigInt(params.teamSupply),
-      vestingStartTime: params.vestingStartTime
-        ? BigInt(params.vestingStartTime)
-        : BigInt(0),
-      vestingDuration: params.vestingDuration
-        ? BigInt(params.vestingDuration)
-        : BigInt(0),
-      vestingWallet: params.vestingWallet
-        ? (params.vestingWallet as Address)
-        : (ethers.ZeroAddress as Address),
-      bondingCurveSupply: BigInt(params.bondingCurveSupply),
-      liquidityPoolSupply: BigInt(params.liquidityPoolSupply),
-      totalSupply: BigInt(totalSupply),
-      bondingCurveBuyFee: BigInt(params.bondingCurveBuyFee),
-      bondingCurveSellFee: BigInt(params.bondingCurveSellFee),
-      bondingCurveFeeSplits: params.bondingCurveFeeSplits.map((split) => ({
-        recipient: split.recipient,
-        bps: split.bps,
-      })),
-      bondingCurveParams: {
-        prices: params.bondingCurveParams.prices.map((price) => BigInt(price)),
-        numSteps: BigInt(params.bondingCurveParams.numSteps),
-        stepSize: BigInt(params.bondingCurveParams.stepSize),
-      },
-      allowForcedGraduation: params.allowForcedGraduation,
-      allowAutoGraduation: params.allowAutoGraduation,
-      graduationFeeBps: BigInt(params.graduationFeeBps),
-      graduationFeeSplits: params.graduationFeeSplits.map((split) => ({
-        recipient: split.recipient,
-        bps: split.bps,
-      })),
-      poolFees: params.poolFees,
-      poolFeeSplits: params.poolFeeSplits.map((split) => ({
-        recipient: split.recipient,
-        bps: split.bps,
-      })),
-      surgeFeeStartingTime: BigInt(Math.floor(Date.now() / 1000)),
-      surgeFeeDuration: BigInt(params.surgeFeeDuration),
-      maxSurgeFeeBps: BigInt(params.maxSurgeFeeBps),
-    };
   }
 }
